+++ packages/playground/src/Select/SelectMultiDemo.tsx
@@ -27,6 +27,7 @@
 import React, { useMemo, useState } from 'react'
 import {
   Box,
+  FieldSelectMulti,
   Heading,
   List,
   ListItem,
@@ -150,6 +151,22 @@ export function SelectMultiDemo() {
 
   return (
     <Box p="large" width={400}>
+      <FieldSelectMulti
+        label="Label"
+        options={selectOptions}
+        placeholder="Search fruits"
+        detail="5/50"
+      />
+      <FieldSelectMulti
+        label="Label"
+        placeholder="placeholder"
+        description="this is the description"
+        options={[
+          { label: 'Cheddar', value: 'cheddar' },
+          { label: 'Gouda', value: 'gouda' },
+          { label: 'Swiss', value: 'swiss' },
+        ]}
+      />
       <Heading mb="large">SelectMulti</Heading>
       <Heading mb="medium" as="h4">
         1k (windowed) options